[STD-282] Continuing entity rework
- Reworking multiple systems
- Reworking resource access (mesh, material, texture)
- Reworking garbage collection
- Reworking entity hierarchy and enabled/disabled entities
- Reworked physics system and api
- Removing textures from decal component

+++ src/renderer/engine/core/resource-libs/CameraResources.ts
@@ -10,7 +10,7 @@ import CameraNotificationDecoder from "../lib/CameraNotificationDecoder"
  * @field projectionBuffer {float32array [zFar, zNear, fov, aR, orthographicSize]}
  */
 export default class CameraResources extends CameraEffects {
-	static position = ArrayBufferAPI.allocateVector(3)
+	static position = ArrayBufferAPI.allocateVector(3) as Float32Array
 	static viewMatrix = ArrayBufferAPI.allocateMatrix(4, true)
 	static projectionMatrix = ArrayBufferAPI.allocateMatrix(4, true)
 	static invViewMatrix = ArrayBufferAPI.allocateMatrix(4, true)
